+++ sbuildr/__init__.py
@@ -3,6 +3,5 @@ from sbuildr.project.profile import Profile
 from sbuildr.tools.flags import BuildFlags
 from sbuildr.tools import compiler, linker
 from sbuildr.graph.node import Library
-from sbuildr.cli.cli import cli
 from sbuildr.logger import G_LOGGER
 __version__ = "0.4.1"